Conflicts:
	src/msgpack/object.h
	src/msgpack/object.hpp
	src/msgpack/pack.hpp
	src/msgpack/type/string.hpp
	src/msgpack/unpack_define.h

+++ src/msgpack/sbuffer.hpp
@@ -47,7 +47,7 @@ public:
 	}
 
 public:
-	void write(const char* buf, unsigned int len)
+	void write(const char* buf, size_t len)
 	{
 		if(alloc_ - size_ < len) {
 			expand_buffer(len);
